Add drag/drop-ability to groups; fix assets bug
Fix a bug from the previous assets conversion (canvas needs to use
"assets" dir, not "static").

For dragging and dropping groups, there was a conundrum with how to
refer to them.  The data structure of order of notes in a dotstorm model
looks like the following -- in this example, <id1> and <id2> are
ungrouped, and <id3> and <id4> are in a group:

    [ <id1>, <id2>, { 'ideas': [<id3>, <id4>] }]

Since groups have no ID, how do we refer to them?  We could do array
index position, but it seemed clearer to use a combination of the ID of
one of their notes, and a boolean indicating whether we were referring
to the group, or the individual note within the group.

This leads to the mildly ridiculous explosion of methods: putLeftOf,
putLeftOfGroup, putGroupLeftOf, and putGroupLeftOfGroup (as well as
'right' for all of those).  I added the combined "move" and "combine"
methods as well (the others alias to the combined methods), but it does
seem clearer to leave the multiple names.