Fix data deletion bug with photo upload
When there was no `photoVersion` attribute of an Idea model, the path
given by `idea.getPhotoPath(size)` was evaluating to the base path --
namely, the root of the assets directory.  Then when we went to delete
the parent directory of the photo, we would clear files in the project
root!!  This is fixed now, with checks in place to ensure that the photo
URL is not null before returning a non-null path.