Fix accidental ungrouping bug, start on trash
Add the placeholder dims to the exclusion criteria for moving drop
targets, so that merely touching a note on the edge of a group won't
cause it to exit the group.

Also, start in on designing trash.  And upgrade jquery -- use unminified
version to permit more meaningful profiling (and we minify everything in
production anyway).