Improve rendering of file input for editor
The opacity hack (http://www.quirksmode.org/dom/inputfile.html) for
styling a file input element only works if the styled input is larger
than the underlying native element.  Otherwise, you get a transparent
click target surrounding the styled input, and potentially conflicting
with other click targets.

Switch instead to this technique
(http://stackoverflow.com/a/3226279/85461), which places the file input
dialog in a separate place, and transmits a click event to it if we
click our styled input.

Note that this strategy will probably not succeed for IE8 etc.