Improve performance for drag and drop
Improve the performance of drag and drop (on mobile especially) by
avoiding touching the DOM whenever possible.

Previously, the drop target interface was:
    onDrag (pos): If the pos matches your drop target, redraw your UI.
    onDrop: If called, execute the drop for this target.
'onDrag' got called on every mouse/touch movement, and thus the same
drop-target UI bits were getting redrawn continuously, even if they
hadn't changed.

Now, the interface is:
    match (pos): If the pos matches your drop target, return true.
    show: Draw your UI element(s).
    hide: Hide your UI element(s).
    onDrop: If called, execute the drop for this target.
Each drop target is free to enclose its current state as to whether it
is matching, and to ignore repeated calls to show() or hide()
accordingly.  The loop for checking drop targets also promises to call
show() or hide() for each drop target on each movement, based on a
combination of the result from match() and the rules for precedence
(only the first matching drop target gets `show()`).

Also: change trash drag-outs to default to the end of the board rather
than beginning.