Improve drag/drop performance on mobile
Rather than re-calculating the positions of all drag/drop targets every
time you pick up a note, do it on resize or when rendering groups.  This
makes drag/drop a snap.

Also, hard-code the width of the dropline, as it doesn't get a correct
size if it hasn't been rendered yet.