Re-organize frontmatter and links
Move the add and trash icons to the header.  Move the dotstorm embed
link to the user menu.  Trash now opens as an absolutely positioned div
over other notes. Tweak drop-target rules so that drags that are within
the trash window, but also over other drop targets, don't trigger the
other drop targets.