Switch to a up/down tally model for voting
Rather than having a vote be a modal switch per session, let voting just
be an arbitrary up/down count.  To avoid accidental multi-voting, add a
2-second timeout between vote presses before you're allowed to vote
again.

This is to allow voting in a scenario where one machine is shared by
multiple people, or where you want to allow individuals to vote multiple
times.  This requires a fairly high degree of group trust, or a fairly
low weight given to the vote tally.